requirements.txt
# Core web/API
fastapi>=0.110
uvicorn[standard]>=0.27

# Config & data
pydantic>=2.6
pydantic-settings>=2.2
orjson>=3.10.0
requests>=2.32.0
httpx>=0.27.0
tenacity>=8.2.0

# Data parsing & cleaning
pandas>=2.2.0
numpy>=1.26.0
lxml>=5.2.0
trafilatura>=1.8.0
pdfminer.six>=20231228
PyPDF2>=3.0.1
beautifulsoup4>=4.12.0

# NLP / retrieval
transformers>=4.40.0
accelerate>=0.29.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
datasketch>=1.6.5
spacy>=3.7.0
jsonschema>=4.22.0

# Optional fine-tuning
peft>=0.11.0
trl>=0.9.4
bitsandbytes>=0.43.0; platform_system == "Linux"

# Excel export
openpyxl>=3.1.2
xlsxwriter>=3.2.0

# Dev tools
pytest>=8.1.0
pytest-cov>=5.0.0
mypy>=1.10.0
ruff>=0.4.7
black>=24.4.0
isort>=5.13.2
pre-commit>=3.7.0
pip-tools>=7.4.0
